Use vmfest hardware DSL.
Add the possibility of using either bridged or local networking.

Add possibility of defining the hardware configuration of the VMs in pallet (via
vmfest hardware DSL).

Provide sane defaults for both networking and hardware.
By default, pallet will provide the best effort to make VMs start successfully.

The metadata of the image is stored in the VM as an extended parameter.

Image selection is based only on predefined selectors matching the ones in
jclouds and will ignore any other entry.  find-images will find all the images
that match a template.

The image metadata is now preprocessed to add some fields automatically for
compatibility with jclouds.